requirements.txt
streamlit==1.39.0
pinecone-client==4.1.2
langchain==0.3.12
langchain-google-genai==2.0.5
python-dotenv==1.0.1
requests
beautifulsoup4
langchain-pinecone
serpapi


